Move PivotPartitioner to a separate package and rename it
