+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/matrixreduction/partitioners/DefaultPivotPartitioner.scala
@@ -1,9 +1,9 @@
-package io.github.jakipatryk.sparkpersistenthomology.matrixreduction
+package io.github.jakipatryk.sparkpersistenthomology.matrixreduction.partitioners
 
 import io.github.jakipatryk.sparkpersistenthomology.Key
 import org.apache.spark.Partitioner
 
-class PivotPartitioner(val numPartitions: Int, val filtrationLength: Long) extends Partitioner {
+class DefaultPivotPartitioner(val numPartitions: Int, val filtrationLength: Long) extends Partitioner {
 
   /**
    * Partitioner used during iteration of PH algorithm.
